fix(ioctl): truncate ioctl request to 32 bits — musl read-direction ioctls (TIOCGPTN) were ENOTTY (#219)
The ioctl handler read the request as the raw 64-bit arg. musl declares ioctl's request param
as `int`, so a read-direction request with the direction bit set (TIOCGPTN=0x80045430) arrives
SIGN-EXTENDED as 0xffffffff80045430 and misses its switch case -> default ENOTTY. glibc declares
it `unsigned long` (zero-extends), which is why python pty.fork()/glibc openpty worked but musl
tmux/script/openpty failed 'Not a tty'. Truncate to 32 bits ((uint32_t)a1) matching the Linux
kernel (cmd is unsigned int) so both forms match — also fixes any other high-direction-bit ioctl
(TCGETS2 etc.) for musl. tmux new-session works both arches; glibc openpty unaffected; ioctl/
getdents/fifo basics green. (Extracted as 1 line onto current HEAD — worktree base predated
#199/#192/#214/#216, all preserved.)

+++ dd-jit/src/runtime/os/linux/syscall/fs.c
@@ -111,7 +111,11 @@ static int svc_fs(struct cpu *c, uint64_t nr, uint64_t a0, uint64_t a1, uint64_t
     // ioctl(fd, req, arg) -- Linux req# -> macOS
     case 29: {
         int fd = (int)a0;
-        unsigned long rq = (unsigned long)a1;
+        // Truncate the ioctl request to 32 bits (Linux `cmd` is unsigned int). musl declares ioctl's request
+        // as `int`, so a read-direction request with the direction bit set (e.g. TIOCGPTN 0x80045430) arrives
+        // SIGN-EXTENDED as 0xffffffff80045430 and would miss its switch case -> ENOTTY (glibc zero-extends, so
+        // it worked there). This makes both forms match, fixing musl tmux/script/openpty and any high-bit ioctl. (#219)
+        unsigned long rq = (uint32_t)a1;
         void *arg = (void *)a2;
         switch (rq) {
         case 0x5401: {
